medic: fix to compile on OSX 10.8

+++ programs/code_medic/code/GDBVarTreeScanner.h
@@ -37,7 +37,7 @@ public:
 
 protected:
 
-	virtual int LexerInput( char* buf, int max_size );
+	virtual size_t LexerInput(char* buf, size_t max_size);
 
 private:
